refactor(operator_commands_ooda/daemon): split tests to daemon_inline.rs (#1266)
- src/operator_commands_ooda/daemon.rs: 643 → 479 LOC
- New tests/daemon_inline.rs (164 LOC), registered in tests/mod.rs
- Widened daemon_log / exe_mtime / interruptible_sleep to pub(super)

Pure refactor: 18 tests pass.
